Moved things around so relevant components are easier to locate Some minor function renaming
