+++ Outset/Functions/FileUtils.swift
@@ -6,49 +6,6 @@
 //
 
 import Foundation
-import CommonCrypto
-
-func runShellCommand(_ command: String, args: [String] = [], verbose: Bool = false) -> (output: String, error: String, exitCode: Int32) {
-    // runs a shell command passed as an argument
-    // If the verbose parameter is set to true, will log the command being run and its status when completed.
-    // returns the output, error and exit code as a tuple.
-
-    if verbose {
-        writeLog("Running task \(command)", logLevel: .debug)
-    }
-    let task = Process()
-    let pipe = Pipe()
-    let errorpipe = Pipe()
-
-    var cmd = command
-    for arg in args {
-        cmd += " '\(arg)'"
-    }
-    let arguments = ["-c", cmd]
-
-    var output: String = ""
-    var error: String = ""
-
-    task.launchPath = "/bin/sh"
-    task.arguments = arguments
-    task.standardOutput = pipe
-    task.standardError = errorpipe
-    task.launch()
-
-    let data = pipe.fileHandleForReading.readDataToEndOfFile()
-    let errordata = errorpipe.fileHandleForReading.readDataToEndOfFile()
-
-    output.append(String(data: data, encoding: .utf8)!)
-    error.append(String(data: errordata, encoding: .utf8)!)
-
-    task.waitUntilExit()
-    let status = task.terminationStatus
-    if verbose {
-        writeLog("Completed task \(command) with status \(status)", logLevel: .debug)
-        writeLog("Task output: \n\(output)", logLevel: .debug)
-    }
-    return (output, error, status)
-}
 
 func installPackage(pkg: String) -> Bool {
     // Installs pkg onto boot drive
@@ -108,87 +65,6 @@ func ensureWorkingFolders() {
             writeLog("could not create path at \(directory)", logLevel: .error)
         }
     }
-}
-
-func migrateLegacyPreferences() {
-    let newoldRootUserDefaults = "/var/root/Library/Preferences/io.macadmins.Outset.plist"
-    // shared folder should not contain any executable content, iterate and update as required
-    if checkFileExists(path: shareDirectory) || checkFileExists(path: newoldRootUserDefaults) {
-        writeLog("Legacy preferences exist. Migrating to user defaults", logLevel: .debug)
-
-        let legacyOutsetPreferencesFile = "\(shareDirectory)com.chilcote.outset.plist"
-        let legacyRootRunOncePlistFile = "com.github.outset.once.\(getConsoleUserInfo().userID).plist"
-        let userHomeDirectory = FileManager.default.homeDirectoryForCurrentUser
-        let userHomePath = userHomeDirectory.relativeString.replacingOccurrences(of: "file://", with: "")
-        let legacyUserRunOncePlistFile = userHomePath+"Library/Preferences/com.github.outset.once.plist"
-
-        var shareFiles: [String] = []
-        shareFiles.append(legacyOutsetPreferencesFile)
-        shareFiles.append(legacyRootRunOncePlistFile)
-        shareFiles.append(legacyUserRunOncePlistFile)
-        shareFiles.append(newoldRootUserDefaults)
-
-        for filename in shareFiles where checkFileExists(path: filename) {
-
-            let url = URL(fileURLWithPath: filename)
-            do {
-                let data = try Data(contentsOf: url)
-                switch filename {
-
-                case newoldRootUserDefaults:
-                    if isRoot() {
-                        writeLog("\(newoldRootUserDefaults) migration", logLevel: .debug)
-                        let legacyDefaultKeys = CFPreferencesCopyKeyList(Bundle.main.bundleIdentifier! as CFString, kCFPreferencesCurrentUser, kCFPreferencesAnyHost)
-                        for key in legacyDefaultKeys as! [CFString] {
-                            let keyValue = CFPreferencesCopyValue(key, Bundle.main.bundleIdentifier! as CFString, kCFPreferencesCurrentUser, kCFPreferencesAnyHost)
-                            CFPreferencesSetValue(key as CFString,
-                                                  keyValue as CFPropertyList,
-                                                  Bundle.main.bundleIdentifier! as CFString,
-                                                  kCFPreferencesAnyUser,
-                                                  kCFPreferencesAnyHost)
-                        }
-                        deletePath(newoldRootUserDefaults)
-                    }
-                case legacyOutsetPreferencesFile:
-                    writeLog("\(legacyOutsetPreferencesFile) migration", logLevel: .debug)
-                    do {
-                        let legacyPreferences = try PropertyListDecoder().decode(OutsetPreferences.self, from: data)
-                        writePreferences(prefs: legacyPreferences)
-                        writeLog("Migrated Legacy Outset Preferences", logLevel: .debug)
-                        deletePath(legacyOutsetPreferencesFile)
-                    } catch {
-                        writeLog("legacy Preferences migration failed", logLevel: .error)
-                    }
-
-                case legacyRootRunOncePlistFile, legacyUserRunOncePlistFile:
-                    writeLog("\(legacyRootRunOncePlistFile) and \(legacyUserRunOncePlistFile) migration", logLevel: .debug)
-                    do {
-                        let legacyRunOncePlistData = try PropertyListDecoder().decode([String: Date].self, from: data)
-                        writeRunOnce(runOnceData: legacyRunOncePlistData)
-                        writeLog("Migrated Legacy Runonce Data", logLevel: .debug)
-                        if isRoot() {
-                            deletePath(legacyRootRunOncePlistFile)
-                        } else {
-                            deletePath(legacyUserRunOncePlistFile)
-                        }
-                    } catch {
-                        writeLog("legacy Run Once Plist migration failed", logLevel: .error)
-                    }
-
-                default:
-                    continue
-                }
-            } catch {
-                writeLog("could not load \(filename)", logLevel: .error)
-            }
-
-        }
-
-        if checkFileExists(path: shareDirectory) && folderContents(path: shareDirectory).isEmpty {
-            deletePath(shareDirectory)
-        }
-    }
-
 }
 
 func checkFileExists(path: String) -> Bool {
@@ -315,97 +191,4 @@ func detachDmg(dmgMount: String) -> String {
         return error
     }
     return output.trimmingCharacters(in: .whitespacesAndNewlines)
-}
-
-func verifySHASUMForFile(filename: String, shasumArray: [String: String]) -> Bool {
-    // Verify that the file
-    var proceed = false
-        let errorMessage = "no required hash or file hash mismatch for: \(filename). Skipping"
-        writeLog("checking hash for \(filename)", logLevel: .debug)
-        let url = URL(fileURLWithPath: filename)
-        if let fileHash = sha256(for: url) {
-            writeLog("file hash : \(fileHash)", logLevel: .debug)
-            if let storedHash = getValueForKey(filename, inArray: shasumArray) {
-                writeLog("required hash : \(storedHash)", logLevel: .debug)
-                if storedHash == fileHash {
-                    proceed = true
-                }
-            }
-        }
-        if !proceed {
-            writeLog(errorMessage, logLevel: .error)
-        }
-
-        return proceed
-}
-
-func sha256(for url: URL) -> String? {
-    // computes a sha256sum for the specified file path and returns a string
-    do {
-        let fileData = try Data(contentsOf: url)
-        let sha256 = fileData.sha256()
-        return sha256.hexEncodedString()
-    } catch {
-        return nil
-    }
-}
-
-func checksumAllFiles() {
-    // compute checksum (SHA256) for all files in the outset directory
-    // returns data in two formats to stdout:
-    //   plaintext
-    //   as plist format ready for import into an MDM or converting to a .mobileconfig
-
-    let url = URL(fileURLWithPath: outsetDirectory)
-    writeLog("CHECKSUM", logLevel: .info)
-    var shasumPlist = FileHashes()
-    if let enumerator = FileManager.default.enumerator(at: url, includingPropertiesForKeys: [.isRegularFileKey], options: [.skipsHiddenFiles, .skipsPackageDescendants]) {
-        for case let fileURL as URL in enumerator {
-            do {
-                let fileAttributes = try fileURL.resourceValues(forKeys: [.isRegularFileKey])
-                if fileAttributes.isRegularFile! && fileURL.pathExtension != "plist" && fileURL.lastPathComponent != "outset" {
-                    if let shasum = sha256(for: fileURL) {
-                        print("\(fileURL.relativePath) : \(shasum)")
-                        shasumPlist.sha256sum[fileURL.relativePath] = shasum
-                    }
-                }
-            } catch { print(error, fileURL) }
-        }
-
-        writeLog("PLIST", logLevel: .info)
-        let encoder = PropertyListEncoder()
-        encoder.outputFormat = .xml
-        do {
-            let data = try encoder.encode(shasumPlist)
-            if let plist = try PropertyListSerialization.propertyList(from: data, options: [], format: nil) as? [String: Any] {
-                let formatted = try PropertyListSerialization.data(fromPropertyList: plist, format: .xml, options: 0)
-                if let string = String(data: formatted, encoding: .utf8) {
-                    print(string)
-                }
-            }
-        } catch {
-            writeLog("plist encoding failed", logLevel: .error)
-        }
-    }
-}
-
-extension Data {
-    // extension to the Data class that lets us compute sha256
-    func sha256() -> Data {
-        var hash = [UInt8](repeating: 0, count: Int(CC_SHA256_DIGEST_LENGTH))
-        self.withUnsafeBytes {
-            _ = CC_SHA256($0.baseAddress, CC_LONG(count), &hash)
-        }
-        return Data(hash)
-    }
-
-    func hexEncodedString() -> String {
-        return map { String(format: "%02hhx", $0) }.joined()
-    }
-}
-
-extension URL {
-    var isDirectory: Bool {
-       (try? resourceValues(forKeys: [.isDirectoryKey]))?.isDirectory == true
-    }
 }
