+++ Outset/Functions/SystemInfo.swift
@@ -1,0 +1,58 @@
+//
+//  SystemInfo.swift
+//  Outset
+//
+//  Created by Bart E Reardon on 5/9/2023.
+//
+
+import Foundation
+
+func getOSVersion() -> String {
+    // Returns the OS version
+    let osVersion = ProcessInfo().operatingSystemVersion
+    let version = "\(osVersion.majorVersion).\(osVersion.minorVersion).\(osVersion.patchVersion)"
+    return version
+}
+
+func getOSBuildVersion() -> String {
+    // Returns the current OS build from sysctl
+    var size = 0
+    sysctlbyname("kern.osversion", nil, &size, nil, 0)
+    var osversion = [CChar](repeating: 0, count: size)
+    sysctlbyname("kern.osversion", &osversion, &size, nil, 0)
+    return String(cString: osversion)
+
+}
+
+func getDeviceSerialNumber() -> String {
+    // Returns the current devices serial number
+    let platformExpert = IOServiceGetMatchingService(kIOMasterPortDefault, IOServiceMatching("IOPlatformExpertDevice") )
+      guard platformExpert > 0 else {
+        return "Serial Unknown"
+      }
+      guard let serialNumber = (IORegistryEntryCreateCFProperty(platformExpert, kIOPlatformSerialNumberKey as CFString, kCFAllocatorDefault, 0).takeUnretainedValue() as? String)?.trimmingCharacters(in: CharacterSet.whitespacesAndNewlines) else {
+        return "Serial Unknown"
+      }
+      IOObjectRelease(platformExpert)
+      return serialNumber
+}
+
+func getMarketingModel() -> String {
+    let appleSiliconProduct = IORegistryEntryFromPath(kIOMasterPortDefault, "IOService:/AppleARMPE/product")
+        let cfKeyValue = IORegistryEntryCreateCFProperty(appleSiliconProduct, "product-description" as CFString, kCFAllocatorDefault, 0)
+        IOObjectRelease(appleSiliconProduct)
+        let keyValue: AnyObject? = cfKeyValue?.takeUnretainedValue()
+        if keyValue != nil, let data = keyValue as? Data {
+            return String(data: data, encoding: String.Encoding.utf8)?.trimmingCharacters(in: CharacterSet(["\0"])) ?? ""
+        }
+        return ""
+}
+
+func getDeviceHardwareModel() -> String {
+    // Returns the current devices hardware model from sysctl
+    var size = 0
+    sysctlbyname("hw.model", nil, &size, nil, 0)
+    var model = [CChar](repeating: 0, count: size)
+    sysctlbyname("hw.model", &model, &size, nil, 0)
+    return String(cString: model)
+}
